+++ channeldb/graph.go
@@ -4176,512 +4176,3 @@ func deserializeLightningNode(r io.Reader) (LightningNode, error) {
 
 	return node, nil
 }
-
-func putChanEdgeInfo(edgeIndex kvdb.RwBucket,
-	edgeInfo *models.ChannelEdgeInfo1, chanID [8]byte) error {
-
-	var b bytes.Buffer
-
-	if _, err := b.Write(edgeInfo.NodeKey1Bytes[:]); err != nil {
-		return err
-	}
-	if _, err := b.Write(edgeInfo.NodeKey2Bytes[:]); err != nil {
-		return err
-	}
-	if _, err := b.Write(edgeInfo.BitcoinKey1Bytes[:]); err != nil {
-		return err
-	}
-	if _, err := b.Write(edgeInfo.BitcoinKey2Bytes[:]); err != nil {
-		return err
-	}
-
-	if err := wire.WriteVarBytes(&b, 0, edgeInfo.Features); err != nil {
-		return err
-	}
-
-	authProof := edgeInfo.AuthProof
-	var nodeSig1, nodeSig2, bitcoinSig1, bitcoinSig2 []byte
-	if authProof != nil {
-		nodeSig1 = authProof.NodeSig1Bytes
-		nodeSig2 = authProof.NodeSig2Bytes
-		bitcoinSig1 = authProof.BitcoinSig1Bytes
-		bitcoinSig2 = authProof.BitcoinSig2Bytes
-	}
-
-	if err := wire.WriteVarBytes(&b, 0, nodeSig1); err != nil {
-		return err
-	}
-	if err := wire.WriteVarBytes(&b, 0, nodeSig2); err != nil {
-		return err
-	}
-	if err := wire.WriteVarBytes(&b, 0, bitcoinSig1); err != nil {
-		return err
-	}
-	if err := wire.WriteVarBytes(&b, 0, bitcoinSig2); err != nil {
-		return err
-	}
-
-	if err := writeOutpoint(&b, &edgeInfo.ChannelPoint); err != nil {
-		return err
-	}
-	if err := binary.Write(&b, byteOrder, uint64(edgeInfo.Capacity)); err != nil {
-		return err
-	}
-	if _, err := b.Write(chanID[:]); err != nil {
-		return err
-	}
-	if _, err := b.Write(edgeInfo.ChainHash[:]); err != nil {
-		return err
-	}
-
-	if len(edgeInfo.ExtraOpaqueData) > MaxAllowedExtraOpaqueBytes {
-		return ErrTooManyExtraOpaqueBytes(len(edgeInfo.ExtraOpaqueData))
-	}
-	err := wire.WriteVarBytes(&b, 0, edgeInfo.ExtraOpaqueData)
-	if err != nil {
-		return err
-	}
-
-	return edgeIndex.Put(chanID[:], b.Bytes())
-}
-
-func fetchChanEdgeInfo(edgeIndex kvdb.RBucket,
-	chanID []byte) (models.ChannelEdgeInfo1, error) {
-
-	edgeInfoBytes := edgeIndex.Get(chanID)
-	if edgeInfoBytes == nil {
-		return models.ChannelEdgeInfo1{}, ErrEdgeNotFound
-	}
-
-	edgeInfoReader := bytes.NewReader(edgeInfoBytes)
-	return deserializeChanEdgeInfo(edgeInfoReader)
-}
-
-func deserializeChanEdgeInfo(r io.Reader) (models.ChannelEdgeInfo1, error) {
-	var (
-		err      error
-		edgeInfo models.ChannelEdgeInfo1
-	)
-
-	if _, err := io.ReadFull(r, edgeInfo.NodeKey1Bytes[:]); err != nil {
-		return models.ChannelEdgeInfo1{}, err
-	}
-	if _, err := io.ReadFull(r, edgeInfo.NodeKey2Bytes[:]); err != nil {
-		return models.ChannelEdgeInfo1{}, err
-	}
-	if _, err := io.ReadFull(r, edgeInfo.BitcoinKey1Bytes[:]); err != nil {
-		return models.ChannelEdgeInfo1{}, err
-	}
-	if _, err := io.ReadFull(r, edgeInfo.BitcoinKey2Bytes[:]); err != nil {
-		return models.ChannelEdgeInfo1{}, err
-	}
-
-	edgeInfo.Features, err = wire.ReadVarBytes(r, 0, 900, "features")
-	if err != nil {
-		return models.ChannelEdgeInfo1{}, err
-	}
-
-	proof := &models.ChannelAuthProof1{}
-
-	proof.NodeSig1Bytes, err = wire.ReadVarBytes(r, 0, 80, "sigs")
-	if err != nil {
-		return models.ChannelEdgeInfo1{}, err
-	}
-	proof.NodeSig2Bytes, err = wire.ReadVarBytes(r, 0, 80, "sigs")
-	if err != nil {
-		return models.ChannelEdgeInfo1{}, err
-	}
-	proof.BitcoinSig1Bytes, err = wire.ReadVarBytes(r, 0, 80, "sigs")
-	if err != nil {
-		return models.ChannelEdgeInfo1{}, err
-	}
-	proof.BitcoinSig2Bytes, err = wire.ReadVarBytes(r, 0, 80, "sigs")
-	if err != nil {
-		return models.ChannelEdgeInfo1{}, err
-	}
-
-	if !proof.IsEmpty() {
-		edgeInfo.AuthProof = proof
-	}
-
-	edgeInfo.ChannelPoint = wire.OutPoint{}
-	if err := readOutpoint(r, &edgeInfo.ChannelPoint); err != nil {
-		return models.ChannelEdgeInfo1{}, err
-	}
-	if err := binary.Read(r, byteOrder, &edgeInfo.Capacity); err != nil {
-		return models.ChannelEdgeInfo1{}, err
-	}
-	if err := binary.Read(r, byteOrder, &edgeInfo.ChannelID); err != nil {
-		return models.ChannelEdgeInfo1{}, err
-	}
-
-	if _, err := io.ReadFull(r, edgeInfo.ChainHash[:]); err != nil {
-		return models.ChannelEdgeInfo1{}, err
-	}
-
-	// We'll try and see if there are any opaque bytes left, if not, then
-	// we'll ignore the EOF error and return the edge as is.
-	edgeInfo.ExtraOpaqueData, err = wire.ReadVarBytes(
-		r, 0, MaxAllowedExtraOpaqueBytes, "blob",
-	)
-	switch {
-	case err == io.ErrUnexpectedEOF:
-	case err == io.EOF:
-	case err != nil:
-		return models.ChannelEdgeInfo1{}, err
-	}
-
-	return edgeInfo, nil
-}
-
-func putChanEdgePolicy(edges kvdb.RwBucket, edge *models.ChannelEdgePolicy1,
-	from, to []byte) error {
-
-	var edgeKey [33 + 8]byte
-	copy(edgeKey[:], from)
-	byteOrder.PutUint64(edgeKey[33:], edge.ChannelID)
-
-	var b bytes.Buffer
-	if err := serializeChanEdgePolicy(&b, edge, to); err != nil {
-		return err
-	}
-
-	// Before we write out the new edge, we'll create a new entry in the
-	// update index in order to keep it fresh.
-	updateUnix := uint64(edge.LastUpdate.Unix())
-	var indexKey [8 + 8]byte
-	byteOrder.PutUint64(indexKey[:8], updateUnix)
-	byteOrder.PutUint64(indexKey[8:], edge.ChannelID)
-
-	updateIndex, err := edges.CreateBucketIfNotExists(edgeUpdateIndexBucket)
-	if err != nil {
-		return err
-	}
-
-	// If there was already an entry for this edge, then we'll need to
-	// delete the old one to ensure we don't leave around any after-images.
-	// An unknown policy value does not have a update time recorded, so
-	// it also does not need to be removed.
-	if edgeBytes := edges.Get(edgeKey[:]); edgeBytes != nil &&
-		!bytes.Equal(edgeBytes[:], unknownPolicy) {
-
-		// In order to delete the old entry, we'll need to obtain the
-		// *prior* update time in order to delete it. To do this, we'll
-		// need to deserialize the existing policy within the database
-		// (now outdated by the new one), and delete its corresponding
-		// entry within the update index. We'll ignore any
-		// ErrEdgePolicyOptionalFieldNotFound error, as we only need
-		// the channel ID and update time to delete the entry.
-		// TODO(halseth): get rid of these invalid policies in a
-		// migration.
-		oldEdgePolicy, err := deserializeChanEdgePolicy(
-			bytes.NewReader(edgeBytes),
-		)
-		if err != nil && err != ErrEdgePolicyOptionalFieldNotFound {
-			return err
-		}
-
-		oldUpdateTime := uint64(oldEdgePolicy.LastUpdate.Unix())
-
-		var oldIndexKey [8 + 8]byte
-		byteOrder.PutUint64(oldIndexKey[:8], oldUpdateTime)
-		byteOrder.PutUint64(oldIndexKey[8:], edge.ChannelID)
-
-		if err := updateIndex.Delete(oldIndexKey[:]); err != nil {
-			return err
-		}
-	}
-
-	if err := updateIndex.Put(indexKey[:], nil); err != nil {
-		return err
-	}
-
-	updateEdgePolicyDisabledIndex(
-		edges, edge.ChannelID,
-		edge.ChannelFlags&lnwire.ChanUpdateDirection > 0,
-		edge.IsDisabled(),
-	)
-
-	return edges.Put(edgeKey[:], b.Bytes()[:])
-}
-
-// updateEdgePolicyDisabledIndex is used to update the disabledEdgePolicyIndex
-// bucket by either add a new disabled ChannelEdgePolicy1 or remove an existing
-// one.
-// The direction represents the direction of the edge and disabled is used for
-// deciding whether to remove or add an entry to the bucket.
-// In general a channel is disabled if two entries for the same chanID exist
-// in this bucket.
-// Maintaining the bucket this way allows a fast retrieval of disabled
-// channels, for example when prune is needed.
-func updateEdgePolicyDisabledIndex(edges kvdb.RwBucket, chanID uint64,
-	direction bool, disabled bool) error {
-
-	var disabledEdgeKey [8 + 1]byte
-	byteOrder.PutUint64(disabledEdgeKey[0:], chanID)
-	if direction {
-		disabledEdgeKey[8] = 1
-	}
-
-	disabledEdgePolicyIndex, err := edges.CreateBucketIfNotExists(
-		disabledEdgePolicyBucket,
-	)
-	if err != nil {
-		return err
-	}
-
-	if disabled {
-		return disabledEdgePolicyIndex.Put(disabledEdgeKey[:], []byte{})
-	}
-
-	return disabledEdgePolicyIndex.Delete(disabledEdgeKey[:])
-}
-
-// putChanEdgePolicyUnknown marks the edge policy as unknown
-// in the edges bucket.
-func putChanEdgePolicyUnknown(edges kvdb.RwBucket, channelID uint64,
-	from []byte) error {
-
-	var edgeKey [33 + 8]byte
-	copy(edgeKey[:], from)
-	byteOrder.PutUint64(edgeKey[33:], channelID)
-
-	if edges.Get(edgeKey[:]) != nil {
-		return fmt.Errorf("cannot write unknown policy for channel %v "+
-			" when there is already a policy present", channelID)
-	}
-
-	return edges.Put(edgeKey[:], unknownPolicy)
-}
-
-func fetchChanEdgePolicy(edges kvdb.RBucket, chanID []byte,
-	nodePub []byte) (*models.ChannelEdgePolicy1, error) {
-
-	var edgeKey [33 + 8]byte
-	copy(edgeKey[:], nodePub)
-	copy(edgeKey[33:], chanID[:])
-
-	edgeBytes := edges.Get(edgeKey[:])
-	if edgeBytes == nil {
-		return nil, ErrEdgeNotFound
-	}
-
-	// No need to deserialize unknown policy.
-	if bytes.Equal(edgeBytes[:], unknownPolicy) {
-		return nil, nil
-	}
-
-	edgeReader := bytes.NewReader(edgeBytes)
-
-	ep, err := deserializeChanEdgePolicy(edgeReader)
-	switch {
-	// If the db policy was missing an expected optional field, we return
-	// nil as if the policy was unknown.
-	case err == ErrEdgePolicyOptionalFieldNotFound:
-		return nil, nil
-
-	case err != nil:
-		return nil, err
-	}
-
-	return ep, nil
-}
-
-func fetchChanEdgePolicies(edgeIndex kvdb.RBucket, edges kvdb.RBucket,
-	chanID []byte) (*models.ChannelEdgePolicy1, *models.ChannelEdgePolicy1,
-	error) {
-
-	edgeInfo := edgeIndex.Get(chanID)
-	if edgeInfo == nil {
-		return nil, nil, fmt.Errorf("%w: chanID=%x", ErrEdgeNotFound,
-			chanID)
-	}
-
-	// The first node is contained within the first half of the edge
-	// information. We only propagate the error here and below if it's
-	// something other than edge non-existence.
-	node1Pub := edgeInfo[:33]
-	edge1, err := fetchChanEdgePolicy(edges, chanID, node1Pub)
-	if err != nil {
-		return nil, nil, fmt.Errorf("%w: node1Pub=%x", ErrEdgeNotFound,
-			node1Pub)
-	}
-
-	// Similarly, the second node is contained within the latter
-	// half of the edge information.
-	node2Pub := edgeInfo[33:66]
-	edge2, err := fetchChanEdgePolicy(edges, chanID, node2Pub)
-	if err != nil {
-		return nil, nil, fmt.Errorf("%w: node2Pub=%x", ErrEdgeNotFound,
-			node2Pub)
-	}
-
-	return edge1, edge2, nil
-}
-
-func serializeChanEdgePolicy(w io.Writer, edge *models.ChannelEdgePolicy1,
-	to []byte) error {
-
-	err := wire.WriteVarBytes(w, 0, edge.SigBytes)
-	if err != nil {
-		return err
-	}
-
-	if err := binary.Write(w, byteOrder, edge.ChannelID); err != nil {
-		return err
-	}
-
-	var scratch [8]byte
-	updateUnix := uint64(edge.LastUpdate.Unix())
-	byteOrder.PutUint64(scratch[:], updateUnix)
-	if _, err := w.Write(scratch[:]); err != nil {
-		return err
-	}
-
-	if err := binary.Write(w, byteOrder, edge.MessageFlags); err != nil {
-		return err
-	}
-	if err := binary.Write(w, byteOrder, edge.ChannelFlags); err != nil {
-		return err
-	}
-	if err := binary.Write(w, byteOrder, edge.TimeLockDelta); err != nil {
-		return err
-	}
-	if err := binary.Write(w, byteOrder, uint64(edge.MinHTLC)); err != nil {
-		return err
-	}
-	if err := binary.Write(w, byteOrder, uint64(edge.FeeBaseMSat)); err != nil {
-		return err
-	}
-	if err := binary.Write(w, byteOrder, uint64(edge.FeeProportionalMillionths)); err != nil {
-		return err
-	}
-
-	if _, err := w.Write(to); err != nil {
-		return err
-	}
-
-	// If the max_htlc field is present, we write it. To be compatible with
-	// older versions that wasn't aware of this field, we write it as part
-	// of the opaque data.
-	// TODO(halseth): clean up when moving to TLV.
-	var opaqueBuf bytes.Buffer
-	if edge.MessageFlags.HasMaxHtlc() {
-		err := binary.Write(&opaqueBuf, byteOrder, uint64(edge.MaxHTLC))
-		if err != nil {
-			return err
-		}
-	}
-
-	if len(edge.ExtraOpaqueData) > MaxAllowedExtraOpaqueBytes {
-		return ErrTooManyExtraOpaqueBytes(len(edge.ExtraOpaqueData))
-	}
-	if _, err := opaqueBuf.Write(edge.ExtraOpaqueData); err != nil {
-		return err
-	}
-
-	if err := wire.WriteVarBytes(w, 0, opaqueBuf.Bytes()); err != nil {
-		return err
-	}
-	return nil
-}
-
-func deserializeChanEdgePolicy(r io.Reader) (*models.ChannelEdgePolicy1, error) {
-	// Deserialize the policy. Note that in case an optional field is not
-	// found, both an error and a populated policy object are returned.
-	edge, deserializeErr := deserializeChanEdgePolicyRaw(r)
-	if deserializeErr != nil &&
-		deserializeErr != ErrEdgePolicyOptionalFieldNotFound {
-
-		return nil, deserializeErr
-	}
-
-	return edge, deserializeErr
-}
-
-func deserializeChanEdgePolicyRaw(r io.Reader) (*models.ChannelEdgePolicy1,
-	error) {
-
-	edge := &models.ChannelEdgePolicy1{}
-
-	var err error
-	edge.SigBytes, err = wire.ReadVarBytes(r, 0, 80, "sig")
-	if err != nil {
-		return nil, err
-	}
-
-	if err := binary.Read(r, byteOrder, &edge.ChannelID); err != nil {
-		return nil, err
-	}
-
-	var scratch [8]byte
-	if _, err := r.Read(scratch[:]); err != nil {
-		return nil, err
-	}
-	unix := int64(byteOrder.Uint64(scratch[:]))
-	edge.LastUpdate = time.Unix(unix, 0)
-
-	if err := binary.Read(r, byteOrder, &edge.MessageFlags); err != nil {
-		return nil, err
-	}
-	if err := binary.Read(r, byteOrder, &edge.ChannelFlags); err != nil {
-		return nil, err
-	}
-	if err := binary.Read(r, byteOrder, &edge.TimeLockDelta); err != nil {
-		return nil, err
-	}
-
-	var n uint64
-	if err := binary.Read(r, byteOrder, &n); err != nil {
-		return nil, err
-	}
-	edge.MinHTLC = lnwire.MilliSatoshi(n)
-
-	if err := binary.Read(r, byteOrder, &n); err != nil {
-		return nil, err
-	}
-	edge.FeeBaseMSat = lnwire.MilliSatoshi(n)
-
-	if err := binary.Read(r, byteOrder, &n); err != nil {
-		return nil, err
-	}
-	edge.FeeProportionalMillionths = lnwire.MilliSatoshi(n)
-
-	if _, err := r.Read(edge.ToNode[:]); err != nil {
-		return nil, err
-	}
-
-	// We'll try and see if there are any opaque bytes left, if not, then
-	// we'll ignore the EOF error and return the edge as is.
-	edge.ExtraOpaqueData, err = wire.ReadVarBytes(
-		r, 0, MaxAllowedExtraOpaqueBytes, "blob",
-	)
-	switch {
-	case err == io.ErrUnexpectedEOF:
-	case err == io.EOF:
-	case err != nil:
-		return nil, err
-	}
-
-	// See if optional fields are present.
-	if edge.MessageFlags.HasMaxHtlc() {
-		// The max_htlc field should be at the beginning of the opaque
-		// bytes.
-		opq := edge.ExtraOpaqueData
-
-		// If the max_htlc field is not present, it might be old data
-		// stored before this field was validated. We'll return the
-		// edge along with an error.
-		if len(opq) < 8 {
-			return edge, ErrEdgePolicyOptionalFieldNotFound
-		}
-
-		maxHtlc := byteOrder.Uint64(opq[:8])
-		edge.MaxHTLC = lnwire.MilliSatoshi(maxHtlc)
-
-		// Exclude the parsed field from the rest of the opaque data.
-		edge.ExtraOpaqueData = opq[8:]
-	}
-
-	return edge, nil
-}
